批量添加文件信息

+++ src/main/java/org/ewlive/entity/system/SysFileInfo.java
@@ -52,7 +52,7 @@ public class SysFileInfo extends Base {
 	 * 文件大小
 	 */
 	@TableField("file_size")
-	private Double fileSize;
+	private Long fileSize;
 
 
 	/**
